moved azureclusteridentity handler to the AzureCluster controller
