+++ service/controller/azurecluster/handler/azureclusteridentity/delete.go
@@ -1,0 +1,10 @@
+package azureclusteridentity
+
+import (
+	"context"
+)
+
+// EnsureDeleted is a no-op.
+func (r *Resource) EnsureDeleted(ctx context.Context, obj interface{}) error {
+	return nil
+}
