+++ service/service.go
@@ -37,7 +37,6 @@ import (
 	"github.com/giantswarm/azure-operator/v5/service/controller/azuremachine"
 	"github.com/giantswarm/azure-operator/v5/service/controller/azuremachinepool"
 	"github.com/giantswarm/azure-operator/v5/service/controller/cluster"
-	credentialctrl "github.com/giantswarm/azure-operator/v5/service/controller/credential"
 	"github.com/giantswarm/azure-operator/v5/service/controller/machinepool"
 	"github.com/giantswarm/azure-operator/v5/service/controller/setting"
 	"github.com/giantswarm/azure-operator/v5/service/controller/unhealthynode"
@@ -435,22 +434,6 @@ func New(config Config) (*Service, error) {
 		}
 
 		controllers = append(controllers, clusterController)
-	}
-
-	var credentialController *operatorkitcontroller.Controller
-	{
-		c := credentialctrl.ControllerConfig{
-			K8sClient: k8sClient,
-			Logger:    config.Logger,
-			SentryDSN: sentryDSN,
-		}
-
-		credentialController, err = credentialctrl.NewController(c)
-		if err != nil {
-			return nil, microerror.Mask(err)
-		}
-
-		controllers = append(controllers, credentialController)
 	}
 
 	var machinePoolController *operatorkitcontroller.Controller
